<refactor>: 디렉토리 구조 다시 잡기
