+++ Rolling/src/components/LandingPage/Article.jsx
@@ -1,5 +1,5 @@
-import img01 from '../assets/landing-img01.svg';
-import img02 from '../assets/landing-img02.svg';
+import img01 from '../../assets/landing-img01.svg';
+import img02 from '../../assets/landing-img02.svg';
 
 function Article() {
   return (
