--- Rolling/src/pages/RollingAddPage.jsx
@@ -1,5 +1,0 @@
-function RollingAddPage() {
-  return;
-}
-
-export default RollingAddPage;
